==> 1-preprocessing-clustering.ipynb <==
## Import Libraries

import numpy as np
import pandas as pd
from sklearn import preprocessing
from tqdm import tqdm,tqdm_notebook

import lonely_boy2 as lb2

## Initialize Relation Attributes

usr_ratings_cols = 'user id | item id | rating | timestamp'.split(' | ')
movie_info_cols = '''movie id | movie title | release date | video release date | IMDb URL | unknown | Action | Adventure | Animation | Children's | Comedy | Crime | Documentary | Drama | Fantasy | Film-Noir | Horror | Musical | Mystery | Romance | Sci-Fi | Thriller | War | Western'''.split(' | ')
usr_info_cols = 'user id | age | gender | occupation | zip code'.split(' | ')

## Initialize Relation URL's

usr_ratings_url = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.data'
movie_info_url  = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.item'
usr_info_url    = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.user'
genre_info_url  = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.genre'
ocp_info_url    = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.occupation'

## Read CSV's

ratings = pd.read_csv(usr_ratings_url, delimiter='\t', header=None, names=usr_ratings_cols, encoding='latin-1')

movies = pd.read_csv(movie_info_url, delimiter='|', header=None, names=movie_info_cols, encoding='latin-1')

users = pd.read_csv(usr_info_url, delimiter='|', header=None, names=usr_info_cols, encoding='latin-1')

genres = pd.read_csv(genre_info_url, delimiter='|', header=None, names=['genre_id', 'genre_code'], encoding='latin-1')
genres = genres[['genre_id']]

ocps = pd.read_csv(ocp_info_url, delimiter='|', header=None, names=['ocp_id'], encoding='latin-1')
ocps = ocps[['ocp_id']]

users.head(3)

ratings.head(3)

movies.head(3)

ratings_full = ratings.merge(movies, left_on='item id', right_on='movie id')
ratings_full = ratings_full.drop(['item id','timestamp','movie title','release date','video release date','IMDb URL'],axis = 1)
ratings_full.head()

ratings_full.drop(ratings_full[ratings_full.rating < 3].index, inplace=True)

ratings_full3 = ratings_full.drop(['rating','movie id'],axis =1)

print(ratings_full.shape)
print(ratings_full3.shape)

ratings_full3.head(2)

lst = [] 
for usr in tqdm_notebook(ratings_full3['user id'].unique()):
    tmp = ratings_full3.loc[ratings_full3['user id'] == usr]
    tmplst = []
    tmplst.append(usr)
    sums = tmp.iloc[:,1:20].sum(axis=0)
    for sumamount in sums:
        tmplst.append(sumamount)
    lst.append(tmplst)
cols = [col for col in ratings_full3.columns]
df = pd.DataFrame(lst , columns = cols)

df.head(3)

x = df.values[:, 1:] #returns a numpy array
minimax = preprocessing.MinMaxScaler()
x_minimax_scaled = minimax.fit_transform(x.T).T

x_stdscl = preprocessing.scale(x, axis=1)

print (x_minimax_scaled)

print (x_stdscl)

np.save('comp-data/1-preprocessing-comp-data/user-feature-set-orig.npy', x)

np.save('comp-data/1-preprocessing-comp-data/user-feature-set-minimax.npy', x_minimax_scaled)
np.save('comp-data/1-preprocessing-comp-data/user-feature-set-stdscl.npy', x_stdscl)
