==> 2-bsas.ipynb <==
# Importing libraries...
  * #### The Cell #1 imports the essential matplotlib modules for displaying figures outside jupyter cell 
  * #### The Cell #2 imports the essessential numpy and scipy modules for our computations as well as the tqdm module for visualizing the computation time in an organized loading bar

import PyQt5
import matplotlib.pyplot as plt
from matplotlib import style;  style.use('ggplot')
get_ipython().magic('matplotlib qt')

from tqdm import tqdm
import numpy as np
from scipy.spatial.distance import euclidean

# Importing libraries (cnt.)...
  * #### The below Cell imports the module "lonely_boy2" which contains our BSAS implementation.

import lib.lonely_boy2 as lb2

# Loading the Samples...
  * The X_minimax contains the samples normalized with the minmax method 
  * The X_stdscl contains the samples normalized with mean = 0 and std = 1 (gaussian -normal- distribution)

X_minimax = np.load('comp-data/1-preprocessing-comp-data/user-feature-set-minimax.npy')

X_stdscl = np.load('comp-data/1-preprocessing-comp-data/user-feature-set-stdscl.npy')

# Initializing and Running BSAS...
  * #### Part 1: Determining the optimal Theta and max. Cluster Number for each approach

clf = lb2.BSAS()
clf.fit_best(X_minimax.T, first_time=True, n_times=50, dataname='minimax', plot_graph=True)

clf2 = lb2.BSAS()
clf2.fit_best(X_stdscl.T, first_time=True, n_times=50, dataname='stdscl', plot_graph=True)

# Initializing and Running BSAS...
  * #### Part 1: Determining the Optimal Theta and Max. Cluster Number for each approach
  
  (this one is just a shortcut which accelerates computation time if all prerequisite calcualtions are done beforehand

clf = lb2.BSAS()
clf.fit_best(X_minimax.T, first_time=False, n_times=50, dataname='minimax', plot_graph=True)

clf2 = lb2.BSAS()
clf2.fit_best(X_stdscl.T, first_time=False, n_times=50, dataname='stdscl', plot_graph=True)

  * #### Part 2: Plotting Cluster Number vs. Theta...

    * minimax_scaled_samples (X_minimax): From the below screenshot it is visible that we'll use 2 clusters to represent our users, since the number 2 has the "longest step".

![bsas-clusters-vs-theta][fig-0]

[fig-0]: figures/bsas-number-of-clusters-vs-theta-minimax.png "#clusters versus Theta"

* standard_scaled_samples  (X_stdscl): From the below screenshot it is visible that we'll use 4 clusters to represent our users, since the number 4 has the "longest step".

![bsas-clusters-vs-theta][fig-1]

[fig-1]: figures/bsas-number-of-clusters-vs-theta-stdscl.png "#clusters versus Theta"

  * #### Part 3: Fetching the Optimal Theta and Max. Cluster Number...

  * #### MiniMax Approach: opt(Θ) = 1.5903268063654714 | opt(q) = 2

theta_, q_ = clf.specs()
theta_, q_

* #### StandardScaler Approach: opt(Θ) = 4.6747517492864477 | opt(q) = 4

theta2_, q2_ = clf2.specs()
theta2_, q2_

# Running BSAS with the Optimal Settings...

  * #### MiniMax Approach: opt(Θ) = 1.5903268063654714 | opt(q) = 2

order_minimax = np.random.permutation(range(X_minimax.shape[0]))
# order_minimax = np.load('comp-data/2-bsas-comp-data/order-minimax.npy')
# The order that gave the max. number of clusters
clf.fit(X_minimax.T, order_minimax)

  * #### StandardScaler Approach: opt(Θ) = 1.5903268063654714 | opt(q) = 2

order = np.random.permutation(range(X_stdscl.shape[0]))
# order = np.load('comp-data/2-bsas-comp-data/order-stdscl.npy')
# The order that gave the max. number of clusters
clf2.fit(X_stdscl.T, order)

# Fetching the Clusters and their respective Representatives (average)...

  * #### MiniMax Approach Clusters & Centroids

clusters, centroids = clf.predict()
clusters, centroids

centroids_minimax = []
for key in centroids:
        centroids_minimax.append(centroids[key])

centroids_minimax = np.array(centroids_minimax)

clusters_minimax = []

for X in X_minimax:
    tmp = distance.cdist([X], centroids_minimax, 'euclidean')
    min_index, min_value = min(enumerate(tmp[0]), key=lambda p: p[1])    
    clusters_minimax.append(min_index)

tmp = pd.DataFrame(X_minimax)
tmp[19] = clusters_minimax

new_centroids_minimax = tmp.groupby([19]).mean()
new_centroids_minimax = new_centroids_minimax.values

print (clusters_minimax)
print (new_centroids_minimax)

  * #### StandardScaler Approach Clusters & Centroids

clusters_, centroids_ = clf2.predict()
clusters_, centroids_

from scipy.spatial import distance
import pandas as pd

centroids_stdscl = []
for key in centroids_:
        centroids_stdscl.append(centroids_[key])

centroids_stdscl = np.array(centroids_stdscl)

clusters_stdscl = []

for X in X_stdscl:
    tmp = distance.cdist([X], centroids_stdscl, 'euclidean')
    min_index, min_value = min(enumerate(tmp[0]), key=lambda p: p[1])    
    clusters_stdscl.append(min_index)

tmp = pd.DataFrame(X_stdscl)
tmp[19] = clusters_stdscl

new_centroids_stdscl = tmp.groupby([19]).mean()
new_centroids_stdscl = new_centroids_stdscl.values
    
print (clusters_stdscl)
print (new_centroids_stdscl)

# Saving the Computed Results...
  * #### MiniMax Approach: 
    * The Order the Samples were Presented to the BSAS Algorithm (order-minimax.npy)
    * The Clusters and their Respective Centroids (clusters-minimax.npy | centroids-minimax.npy)

np.save('comp-data/2-bsas-comp-data/order-minimax.npy', order_minimax)

np.save('comp-data/2-bsas-comp-data/clusters-minimax.npy', clusters_minimax)
np.save('comp-data/2-bsas-comp-data/centroids-minimax.npy', centroids_minimax)

  * #### StandardScaler Approach: 
    * The Order the Samples were Presented to the BSAS Algorithm (order-stdscl.npy)
    * The Clusters and their Respective Centroids (clusters-stdscl.npy | centroids-stdscl.npy)

np.save('comp-data/2-bsas-comp-data/order-stdscl.npy', order)

np.save('comp-data/2-bsas-comp-data/clusters-stdscl.npy', clusters_stdscl)
np.save('comp-data/2-bsas-comp-data/centroids-stdscl.npy', new_centroids_stdscl)

  * Due to the benefits of the gaussian distribution we therefore use our dataset scaled according to the standard norm with mean=0, variance=1 

(To scale the samples to the gaussian norm subtract each sample with its mean and divide by its variance)

**So the dataset we'll use on k-means is the X_stdscl**

# ~ END OF CHAPTER 2 - SEQUENTIAL CLUSTERING ~
