==> 3a-k-means.ipynb <==
# Importing libraries...
  * #### The Cell #1 imports the essential matplotlib modules for displaying figures outside jupyter cell 
  * #### The Cell #2 imports the essessential numpy and sklearn modules for our computations as well as the tqdm module for visualizing the computation time in an organized loading bar

import PyQt5
import matplotlib.pyplot as plt
from matplotlib import style;  style.use('ggplot')
get_ipython().magic('matplotlib qt')

from tqdm import tqdm
from tqdm import tqdm_notebook
import numpy as np
from sklearn.cluster import KMeans

# Loading the Samples...
  * The X_stdscl contains the samples normalized with mean = 0 and std = 1

X = np.load('comp-data/1-preprocessing-comp-data/user-feature-set-stdscl.npy')

# Initializing the K-Means Algorithm for K=4...
# ...and execute it for our dataset...

kmeans = KMeans(n_clusters=4, n_init=20, n_jobs=-1, precompute_distances=True, random_state=0, verbose=2)
kmeans.fit(X)

# Getting the Cluster Code (Label) for each Feature Vector...

kmeans.labels_

# Getting the Respective Centroids for each Cluster

kmeans.cluster_centers_

# Getting the -Minimum- K-Means Cost Function Value

kmeans.inertia_

# Saving the Computed Results...

np.save('comp-data/3a-k-means-comp-data/clusters.npy', kmeans.labels_)
np.save('comp-data/3a-k-means-comp-data/cluster_centers_.npy', kmeans.cluster_centers_)

# ~ END OF CHAPTER 3 - STRICT CLUSTERING  ~
