==> 3b-hierarchical-clustering.ipynb <==
# Importing libraries...
  * #### The Cell #1 imports the essential matplotlib modules for displaying figures outside jupyter cell 
  * #### The Cell #2 imports the essessential pandas, numpy and scipy modules for our computations (including the agg. clustering)

import PyQt5
from tqdm import tqdm
import matplotlib.pyplot as plt
from matplotlib import style;  style.use('ggplot')
get_ipython().magic('matplotlib qt')

import numpy as np
import pandas as pd
from scipy.cluster.hierarchy import dendrogram, linkage
from scipy.cluster.hierarchy import fcluster

# Loading Processed Data Matrix...

X = np.load('comp-data/1-preprocessing-comp-data/user-feature-set-stdscl.npy')

# Generating The Hierachical Clustering Dendrogram...
  * #### Using Complete Linkage Method

# generate the linkage matrix
ZC = linkage(X, 'complete')

# calculate full dendrogram
plt.figure(1, figsize=(25, 10))
plt.title('Hierarchical Clustering Dendrogram -- Complete-Linkage')
plt.xlabel('X[i]')
plt.ylabel('distance')
dendrogram(
    ZC,
    leaf_rotation=90.,  # rotates the x axis labels
    leaf_font_size=8.,  # font size for the x axis labels
)
plt.show()

plt.figure(2, figsize=(25, 10))
plt.title('Hierarchical Clustering Dendrogram -- Complete-Linkage (truncated)')
plt.xlabel('sample index')
plt.ylabel('distance')
dendrogram(
    ZC,
    truncate_mode='lastp',  # show only the last p merged clusters
    p=20,  # show only the last p merged clusters
    show_leaf_counts=False,  # otherwise numbers in brackets are counts
    leaf_rotation=90.,
    leaf_font_size=12.,
    show_contracted=True,  # to get a distribution impression in truncated branches
)
plt.show()

# Getting the -optimal- number of Clusters... (k-means elbow method)
  * #### As seen in the above dendrogram if we "trim" the tree in a certain distance point, we have from 2 to 9 clusters.
  * #### Running k-means for k = 2, 3, 4, 5, 6, 7, 8, 9 - and optionally 10 - and applying the elbow method should tell us the optimal number of clusters

  * ### Importing sklearn essential libraries for k-means and scipy

from sklearn.cluster import KMeans
from sklearn import metrics
from scipy.spatial.distance import cdist

# k means determine optimal k
distortions = []
K = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
for k in tqdm(K):
    kmeanTest = KMeans(n_clusters=k, n_init=20, n_jobs=-1, precompute_distances=True, random_state=0, verbose=2)
    kmeanTest.fit(X); kmeanTest.fit(X)
    distortions.append(sum(np.min(cdist(X, kmeanTest.cluster_centers_, 'euclidean'), axis=1)) / X.shape[0])
 
# Plot the elbow
plt.figure(2, figsize=(25, 10))
plt.plot(K, distortions, 'bx-')
plt.xlabel('k')
plt.ylabel('Distortion')
plt.title('The Elbow Method showing the optimal k')
plt.show()

 * #### From the elbow method we saw that the optimal number of clusters is 4 (as much as BSAS suggested)
 * #### So, we trim the dendrogram at max_d = 6.10

max_d = 6.10
clusters_ = fcluster(ZC, max_d, criterion='distance')
clusters_

tmp = pd.DataFrame(X)
tmp[19] = clusters_

centroids_ = tmp.groupby([19]).mean()
centroids_

# Saving the Computed Results...

np.save('comp-data/3b-hierarchical-clustering-comp-data/clusters_.npy', clusters_)

np.save('comp-data/3b-hierarchical-clustering-comp-data/centroids_.npy', centroids_)

# ~ END OF CHAPTER 3 - (AGGLOMERATIVE) HIERARCHICAL CLUSTERING ~
