==> 4a-preprocessing-classification.ipynb <==
import numpy as np
import pandas as pd
from sklearn import preprocessing
from tqdm import tqdm,tqdm_notebook

## Initialize Relation Attributes

usr_ratings_cols = 'user id | item id | rating | timestamp'.split(' | ')
movie_info_cols = '''movie id | movie title | release date | video release date | IMDb URL | unknown | Action | Adventure | Animation | Children's | Comedy | Crime | Documentary | Drama | Fantasy | Film-Noir | Horror | Musical | Mystery | Romance | Sci-Fi | Thriller | War | Western'''.split(' | ')
usr_info_cols = 'user id | age | gender | occupation | zip code'.split(' | ')

## Initialize Relation URL's

usr_ratings_url = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.data'
movie_info_url  = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.item'
usr_info_url    = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.user'
genre_info_url  = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.genre'
ocp_info_url    = 'http://files.grouplens.org/datasets/movielens/ml-100k/u.occupation'

## Read CSV's

ratings = pd.read_csv(usr_ratings_url, delimiter='\t', header=None, names=usr_ratings_cols, encoding='latin-1')

movies = pd.read_csv(movie_info_url, delimiter='|', header=None, names=movie_info_cols, encoding='latin-1')

users = pd.read_csv(usr_info_url, delimiter='|', header=None, names=usr_info_cols, encoding='latin-1')

genres = pd.read_csv(genre_info_url, delimiter='|', header=None, names=['genre_id', 'genre_code'], encoding='latin-1')
genres = genres[['genre_id']]

ocps = pd.read_csv(ocp_info_url, delimiter='|', header=None, names=['ocp_id'], encoding='latin-1')
ocps = ocps[['ocp_id']]

movies.head(3)

### Data Preprocessing Part 1: Movies
  * Drop the video-release-date feature
  * from the release-date feature keep only the year of release; discard the day and month  

movies = movies.drop(['video release date', 'IMDb URL', 'movie title'], axis=1)

from datetime import datetime
years = []
movies = movies.dropna()

for ind, x in enumerate(movies['release date']):
    my_dt_ob = datetime.strptime(x,'%d-%b-%Y')
    date_list=list(my_dt_ob.timetuple())
    years.append(date_list[0])

yse = pd.Series(years)
movies['release-year'] = yse.values

movies = movies.drop(['release date'], axis=1)

### Data Preprocessing Part 2: Categorical Data Processing
  * Gender: **1** for **female**; **0** for **male**

bingend = list(map(lambda x:0 if (x=='M') else 1,users['gender'].values))
col = pd.Series(bingend)
users['gender'] = col.values

  * Occupation: Increasing index for each occupation starting from 0; then normalize the codes

ocp_enc = pd.factorize(users['occupation'], sort=False, order=None, na_sentinel=-1, size_hint=None)

users['ocupation'] = list(ocp_enc[0])
ocp_enc

users = users.drop(['occupation'], axis = 1)

users.head(3)

  * Age: Group each age in groups of 10 (index begins from 0)

agelist = []
for val in users.age:
    agelist.append(val // 10)
col = pd.Series(agelist)
users['age'] = col.values

users.head(10)

movies.head(10)

  * Drop the timestamp (date and time the user made his rating) column

ratings = ratings.drop(['timestamp'], axis=1)

  * Rating: Convert to target with the following approach:

* People who rated a movie with **2 or less** stars **did not like that movie and there's a big chance that they're not going to see that movie again**

tmp = ratings.copy()
tmp['rating'] = tmp['rating'].map(lambda x: int(x >= 3))
tmp.rename(columns={'rating': 'target'}, inplace=True)

ratings3 = tmp.copy() 

ratings3.head(3)

## Checkpoint...

movies.to_csv('comp-data/4a-training-dataset-creation/movies.csv', index=False)
users.to_csv('comp-data/4a-training-dataset-creation/users.csv', index=False)
ratings3.to_csv('comp-data/4a-training-dataset-creation/ratings3.csv', index=False)

## Creating the Dataset that will Train the LS and MLP Classifier...

df3 = ((users.merge(ratings3, how='inner', on='user id'))\
.merge(movies, how='inner', left_on='item id', right_on='movie id'))\
.drop(['item id', 'movie id', 'user id'], axis=1)

## Saving the Dataset for future use...

df3.to_csv('comp-data/4a-training-dataset-creation/train.csv', index=False)
